question before saving updated info

+++ main.py
@@ -36,6 +36,26 @@ else:
 print('\n SORTED DATA BY WEIGHT WITH CALCULATED MISSING VALUES: ')
 print(df.sort_values('Weight', ascending=True).to_string(index=False))
 
-# write updated data into a file
-df.to_csv('updated_data.csv', index=False)
-print('\n\n\n *** Updated data has been saved into updated_data.csv *** \n\n\n')
+
+def ask_user(question):
+    check = str(input("Question ? (Y/N): ")).lower().strip()
+    try:
+        if check[0] == 'y':
+            return True
+        elif check[0] == 'n':
+            return False
+        else:
+            print('Invalid Input')
+            return ask_user()
+    except Exception as error:
+        print("Please enter valid inputs")
+        print(error)
+        return ask_user()
+
+
+if ask_user("Do you want to save/overwrite updated info into a file?"):
+    # write updated data into a file
+    df.to_csv('updated_data.csv', index=False)
+    print('\n\n\n *** Updated data has been saved into updated_data.csv *** \n\n\n')
+else:
+    print('\n\n\n *** Data has not been saved *** \n\n\n')
